Move `FileContents` to `org.gradle.api.file`
